+++ sync-for-commerce/CodatSyncCommerce/Models/Operations/ListIntegrationsResponse.cs
@@ -23,6 +23,11 @@ namespace CodatSyncCommerce.Models.Operations
         public string? ContentType { get; set; } = default!;
 
         /// <summary>
+        /// Your `query` parameter was not correctly formed
+        /// </summary>
+        public ErrorMessage? ErrorMessage { get; set; }
+
+        /// <summary>
         /// Success
         /// </summary>
         public Integrations? Integrations { get; set; }