+++ sync-for-commerce/RELEASES.md
@@ -83,3 +83,13 @@ Based on:
 - [csharp v2.2.0] sync-for-commerce
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Commerce/2.2.0 - sync-for-commerce
+
+## 2023-11-06 01:25:18
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.0] sync-for-commerce
+### Releases
+- [NuGet v2.3.0] https://www.nuget.org/packages/Codat.Sync.Commerce/2.3.0 - sync-for-commerce